refactor: clean up CalculatorPage and update imports
Refactor the CalculatorPage component to code 
readability by adjusting indentation and adding a container 
class for better styling. Update the import statement in 
App.jsx to reflect the correct path for the Components 
module. Remove the unused LandingPage component to 
streamline the codebase and enhance maintainability.
