+++ src/pages/CalculatorPage.jsx
@@ -1,23 +1,21 @@
 function CalculatorPage() {
-  return (
-    <div className="prose">
-      <h2>
-        Calculator Page
-      </h2>
-      <p>
-        Lorem, ipsum dolor sit amet consectetur adipisicing elit. Harum, facilis
-        assumenda! Alias qui, ipsum omnis amet dolore voluptatibus natus,
-        sapiente inventore quo pariatur sit maiores corrupti asperiores aperiam
-        sunt deserunt.
-      </p>
-      <p>
-        Lorem ipsum dolor sit amet, consectetur adipisicing elit. Fugit error
-        officia hic maiores corporis, sequi similique impedit reiciendis tenetur
-        distinctio aliquam delectus ullam vitae quia laudantium incidunt iste,
-        odit officiis.
-      </p>
-    </div>
-  );
+    return (
+        <div className="prose container mx-auto">
+            <h2>Calculator Page</h2>
+            <p>
+                Lorem, ipsum dolor sit amet consectetur adipisicing elit. Harum,
+                facilis assumenda! Alias qui, ipsum omnis amet dolore
+                voluptatibus natus, sapiente inventore quo pariatur sit maiores
+                corrupti asperiores aperiam sunt deserunt.
+            </p>
+            <p>
+                Lorem ipsum dolor sit amet, consectetur adipisicing elit. Fugit
+                error officia hic maiores corporis, sequi similique impedit
+                reiciendis tenetur distinctio aliquam delectus ullam vitae quia
+                laudantium incidunt iste, odit officiis.
+            </p>
+        </div>
+    );
 }
 
 export default CalculatorPage;
