+++ src/pages/ResultsPage.jsx
@@ -1,29 +1,27 @@
 function ResultsPage() {
-  return (
-    <div className="prose">
-      <h2>
-        Results Page
-      </h2>
-      <p>
-        Lorem ipsum dolor, sit amet consectetur adipisicing elit. Dignissimos
-        molestias reiciendis aperiam minima soluta explicabo porro deserunt fuga
-        cumque et iure, suscipit non sed ex nesciunt. Necessitatibus reiciendis
-        sint accusamus.
-      </p>
-      <p>
-        Lorem ipsum dolor sit amet, consectetur adipisicing elit. Fugit error
-        officia hic maiores corporis, sequi similique impedit reiciendis tenetur
-        distinctio aliquam delectus ullam vitae quia laudantium incidunt iste,
-        odit officiis.
-      </p>
-      <p>
-        Lorem ipsum dolor sit amet, consectetur adipisicing elit. Fugit error
-        officia hic maiores corporis, sequi similique impedit reiciendis tenetur
-        distinctio aliquam delectus ullam vitae quia laudantium incidunt iste,
-        odit officiis.
-      </p>
-    </div>
-  );
+    return (
+        <div className="prose container mx-auto">
+            <h2>Results Page</h2>
+            <p>
+                Lorem ipsum dolor, sit amet consectetur adipisicing elit.
+                Dignissimos molestias reiciendis aperiam minima soluta explicabo
+                porro deserunt fuga cumque et iure, suscipit non sed ex
+                nesciunt. Necessitatibus reiciendis sint accusamus.
+            </p>
+            <p>
+                Lorem ipsum dolor sit amet, consectetur adipisicing elit. Fugit
+                error officia hic maiores corporis, sequi similique impedit
+                reiciendis tenetur distinctio aliquam delectus ullam vitae quia
+                laudantium incidunt iste, odit officiis.
+            </p>
+            <p>
+                Lorem ipsum dolor sit amet, consectetur adipisicing elit. Fugit
+                error officia hic maiores corporis, sequi similique impedit
+                reiciendis tenetur distinctio aliquam delectus ullam vitae quia
+                laudantium incidunt iste, odit officiis.
+            </p>
+        </div>
+    );
 }
 
 export default ResultsPage;
